A couple more bitness matching improvements.
* subversion/libsvn_subr/target.c
  (svn_path_condense_targets):
* subversion/libsvn_subr/svn_string.c
  (svn_cstring_join):
  Use a size_t for the result of strlen().


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1166748 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_subr/svn_string.c
@@ -725,7 +725,7 @@ svn_cstring_join(const apr_array_header_t *strings,
                  apr_pool_t *pool)
 {
   svn_stringbuf_t *new_str = svn_stringbuf_create("", pool);
-  int sep_len = strlen(separator);
+  size_t sep_len = strlen(separator);
   int i;
 
   for (i = 0; i < strings->nelts; i++)
